# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "HexSchool-Meta-Wall",
    "description": "六角學院2024Node.js春季班API"
  },
  "servers": [
    {
      "url": "http://localhost:2330",
      "description": "Local"
    },
    {
      "url": "https://hexschool-meta-wall.onrender.com",
      "description": "Render"
    }
  ],
...
  "paths": {
    "/v1/api/upload": {
      "post": {
        "tags": ["圖片上傳"],
        "summary": "圖片上傳",
        "description": "圖片上傳",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "imageUrl": {
                      "type": "string",
                      "example": "image URL(https)"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "imageUrl": {
                      "type": "string",
                      "example": "image URL(https)"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "貼文牆",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "file"
                  }
                },
                "required": ["file"]
              }
            }
          }
        }
      }
    },
...
    "/v1/api/posts": {
      "get": {
        "tags": ["貼文牆管理"],
        "summary": "查詢全部貼文",
        "description": "查詢全部貼文",
        "parameters": [
          {
            "name": "timeSort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyWord",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "true"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "token": {
                              "type": "string",
                              "example": "eyJhbGciOiJ..........mDWPvJZSxu98W4"
                            },
                            "name": {
                              "type": "string",
                              "example": "Lobinda"
                            }
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "true"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "token": {
                              "type": "string",
                              "example": "eyJhbGciOiJ..........mDWPvJZSxu98W4"
                            },
                            "name": {
                              "type": "string",
                              "example": "Lobinda"
                            }
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["貼文牆管理"],
        "summary": "新增貼文",
        "description": "新增貼文",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "已建立貼文"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "string",
                          "example": "665c618fa2685f9b9e4c79fe"
                        },
                        "content": {
                          "type": "string",
                          "example": "這是第一條訊息"
                        },
                        "userPhoto": {
                          "type": "string",
                          "example": "https://unsplash.com/photos/a-black-and-white-photo-of-a-woman-holding-a-cell-phone-3D0Jxa1QSVA"
                        },
                        "likes": {
                          "type": "array",
                          "example": [],
                          "items": {}
                        },
                        "tags": {
                          "type": "array",
                          "example": ["美食"],
                          "items": {
                            "type": "string"
                          }
                        },
                        "image": {
                          "type": "string",
                          "example": "https://unsplash.com/photos/a-black-and-white-photo-of-a-woman-holding-a-cell-phone-3D0Jxa1QSVA"
                        },
                        "_id": {
                          "type": "string",
                          "example": "665c93a755a500762b17d3e6"
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2024-06-02T15:45:43.913Z"
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "已建立貼文"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "string",
                          "example": "665c618fa2685f9b9e4c79fe"
                        },
                        "content": {
                          "type": "string",
                          "example": "這是第一條訊息"
                        },
                        "userPhoto": {
                          "type": "string",
                          "example": "https://unsplash.com/photos/a-black-and-white-photo-of-a-woman-holding-a-cell-phone-3D0Jxa1QSVA"
                        },
                        "likes": {
                          "type": "array",
                          "example": [],
                          "items": {}
                        },
                        "tags": {
                          "type": "array",
                          "example": ["美食"],
                          "items": {
                            "type": "string"
                          }
                        },
                        "image": {
                          "type": "string",
                          "example": "https://unsplash.com/photos/a-black-and-white-photo-of-a-woman-holding-a-cell-phone-3D0Jxa1QSVA"
                        },
                        "_id": {
                          "type": "string",
                          "example": "665c93a755a500762b17d3e6"
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2024-06-02T15:45:43.913Z"
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "貼文牆",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "example": "這是第一條訊息"
                  },
                  "userPhoto": {
                    "type": "string",
                    "example": "https://unsplash.com/photos/a-black-and-white-photo-of-a-woman-holding-a-cell-phone-3D0Jxa1QSVA"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": "美食"
                  },
                  "image": {
                    "type": "string",
                    "example": "https://unsplash.com/photos/a-black-and-white-photo-of-a-woman-holding-a-cell-phone-3D0Jxa1QSVA"
                  }
                },
                "required": ["content"]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}